Moved windows specific component to a seperate project.
