+++ Daz Package Manager/Daz Package Manager.csproj
@@ -5,7 +5,6 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>Daz_Package_Manager</RootNamespace>
     <UseWPF>true</UseWPF>
-	<UseWindowsForms>true</UseWindowsForms>
 	<ApplicationManifest>app.manifest</ApplicationManifest>
 	<StartupObject></StartupObject>
 	<AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -22,6 +21,7 @@
   <ItemGroup>
     <ProjectReference Include="..\DazMetadata\DazPackage.csproj" />
     <ProjectReference Include="..\Helper\Helpers.csproj" />
+    <ProjectReference Include="..\OsHelper\OsHelper.csproj" />
   </ItemGroup>
 
   <ItemGroup>
