+++ Helper/Helpers.csproj
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <UseWPF>true</UseWPF>
-	<UseWindowsForms>true</UseWindowsForms>
 	<Configurations>Debug;Release;Shipping</Configurations>
 
   </PropertyGroup>
